homework: prepare test system for hardcore test cases
As there is a ton of students enrolled at a course,
program breaking MUST be automatized.

Prepare a homemade test system for this purpose.
Also start collecting interesting and unusual examples.

To use the system, put student's code file into a folder and run:
PROGRAM=<program> make

To add a test case either modify a Makefile or put an .input file
into the tests folder.